requirements.txt
tortoise-orm
fastapi
uvicorn[standard]